user filter and delete users

+++ Laba5/Controllers/UserController.cs
@@ -1,9 +1,19 @@
-﻿using Laba4.Models;
+﻿using Laba4.Data.Cache;
+using Laba4.Models;
 using Laba4.ViewModels;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using PostCity.Models;
+using PostCity.ViewModels.Filters;
+using PostCity.ViewModels;
+using Laba4.ViewModels.Filters.FilterModel;
+using System.Net;
+using PostCity.Data.Cookies;
+using PostCity.Infrastructure.Filters;
+using PostCity.ViewModels.Sort;
+using Laba4.ViewModels.Sort;
 
 namespace Laba4.Controllers
 {
@@ -12,49 +22,61 @@ namespace Laba4.Controllers
     {
         private readonly RoleManager<IdentityRole> _roleManager;
         private readonly UserManager<PostCityUser> _userManager;
+        private readonly UserCache _cache;
         private readonly PostCityContext _context;
+        private readonly CookiesManeger _cookies;
+        private readonly FilterBy<UserViewModel> _filter;
 
-        public UserController(RoleManager<IdentityRole> roleManager, UserManager<PostCityUser> userManager, PostCityContext context)
+        public UserController(RoleManager<IdentityRole> roleManager,
+                              UserManager<PostCityUser> userManager, 
+                              PostCityContext context,
+                              UserCache cache,
+                              CookiesManeger cookiesManeger,
+                              FilterBy<UserViewModel> filter)
         {
             _roleManager = roleManager;
             _userManager = userManager;
             _context = context;
+            _cache = cache;
+            _cookies = cookiesManeger;
+            _filter = filter;
         }
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(UserSortState sortOrder = UserSortState.StandardState, int page = 1)
         {
-            var users = await _userManager.Users.ToListAsync(); 
+            var users = _cache.Get();
 
-            var usersList = new List<UserViewModel>();
+            UserFilterModel filterData = _cookies.GetFromCookies<UserFilterModel>(Request.Cookies, "UserFilterData");
 
-            foreach (var user in users)
-            {
-                var roles = await _userManager.GetRolesAsync(user);
 
-                var roleToDisplay = roles.FirstOrDefault();
+            SetSortOrderViewData(sortOrder);
+            users = ApplySortOrder(users, sortOrder);
 
-                IUser userObj = null;
+            int pageSize = 10;
+            _cache.Set(users);
 
-                if (roleToDisplay == "Recipient")
-                {
-                    userObj = await _context.Recipients.FindAsync(user.UserId);
-                }
-                else if (roleToDisplay == "Employee")
-                {
-                    userObj = await _context.Employees.FindAsync(user.UserId);
-                }
-                
+            var pageViewModel = new PageViewModel<UserViewModel, UserFilterModel>(users, page, pageSize, filterData);
+            return View(pageViewModel);
 
-                usersList.Add(new UserViewModel()
-                {
-                    Id = user.Id,
-                    UserObj = userObj,
-                    UserId = user.UserId,
-                    Role = roleToDisplay,
-                    Email = user.Email,
-                });
-            }
+        }
 
-            return View(usersList);
+        [HttpPost]
+        public async Task<IActionResult> Index(UserFilterModel filterData, int page = 1)
+        {
+            _cache.Update();
+            _cookies.SaveToCookies(Response.Cookies, "UserFilterData", filterData);
+
+            var users = _cache.Get();
+
+            users = _filter.FilterByString(users, fn => fn.UserObj != null ? fn.UserObj.FullName : "", filterData.Name).ToList();
+            users = _filter.FilterByString(users, e => e.Email, filterData.Email).ToList();
+            users = _filter.FilterByString(users, r => r.Role, filterData.Role).ToList();
+
+            int pageSize = 10;
+            _cache.Set(users);
+
+            var pageViewModel = new PageViewModel<UserViewModel, UserFilterModel>(users, page, pageSize, filterData);
+            return View(pageViewModel);
+
         }
 
 
@@ -72,6 +94,7 @@ namespace Laba4.Controllers
                 IdentityResult result = await _roleManager.CreateAsync(new IdentityRole(name));
                 if (result.Succeeded)
                 {
+                    _cache.Update();
                     return RedirectToAction("Index");
                 }
                 else
@@ -91,6 +114,7 @@ namespace Laba4.Controllers
             if (role != null)
             {
                 IdentityResult result = await _roleManager.DeleteAsync(role);
+
             }
             return RedirectToAction("Index");
         }
@@ -101,12 +125,57 @@ namespace Laba4.Controllers
             PostCityUser user = await _userManager.FindByIdAsync(id);
             if (user != null)
             {
-                IdentityResult result = await _userManager.DeleteAsync(user);
+                    var roles = await _userManager.GetRolesAsync(user);
+                    var userRole = roles.FirstOrDefault();
+
+                    if (userRole == "Employee")
+                    {
+                        var employee = await _context.Employees.FindAsync(user.UserId);
+                        if (employee != null)
+                        {
+                            _context.Employees.Remove(employee);
+                        }
+                    }
+                    else if(userRole == "Recipient")
+                    {
+                        var recipient = await _context.Recipients.FindAsync(user.UserId);
+                        if (recipient != null)
+                        {
+                            _context.Recipients.Remove(recipient);
+                        }
+                    }
+
+                    int isDelete = await _context.SaveChangesAsync();
+
+                    if(isDelete == 1)
+                    {
+
+                        IdentityResult result = await _userManager.DeleteAsync(user);
+                    }
+                
+                    _cache.Update();
             }
             return RedirectToAction("Index");
         }
 
         public IActionResult Roles() => View(_roleManager.Roles);
 
+        public void SetSortOrderViewData(UserSortState sortOrder)
+        {
+            ViewData["RoleSort"] = sortOrder == UserSortState.RoleAsc
+                ? UserSortState.RoleDesc
+                : UserSortState.RoleAsc;
+        }
+
+        public IEnumerable<UserViewModel> ApplySortOrder(IEnumerable<UserViewModel> postCityContext, UserSortState sortOrder)
+        {
+            return sortOrder switch
+            {
+                UserSortState.RoleAsc => postCityContext.OrderByDescending(r => r.Role),
+                UserSortState.RoleDesc => postCityContext.OrderBy(r => r.Role),
+
+                UserSortState.StandardState => postCityContext,
+            };
+        }
     }
 }
